+++ src/main/java/com/neo/byez/domain/user/UserDto.java
@@ -1,4 +1,4 @@
-package com.neo.byez.domain;
+package com.neo.byez.domain.user;
 
 import java.sql.Timestamp;
