+++ src/main/resources/mapper/user/userInfoHistMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.neo.byez.dao.UserInfoHistDao">
+<mapper namespace="com.neo.byez.dao.user.UserInfoHistDao">
 
     <!-- 테스트용 쿼리문 -->
     <!-- 전체 데이터 삭제 -->
