+++ src/main/resources/mapper/user/userMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.neo.byez.dao.UserDao">
+<mapper namespace="com.neo.byez.dao.user.UserDao">
 
     <!-- 현재 시간 반환하는 SQL 문 -->
     <select id="selectNow" resultType="java.sql.Timestamp">
